fixed the job issue
