+++ frontend/src/contexts/settingsContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const SettingsContext = createContext();
