+++ frontend/src/contexts/taskContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const TaskContext = createContext();
